ref: renamed ProductRight to ProductDetail and changed all references and moved files to ProductImage component
