+++ client/components/ProductDetails/ProductDetails.styles.ts
@@ -1,6 +1,6 @@
 import styled from "styled-components";
 
-export const ProductRightContainer = styled.div`
+export const ProductDetailsContainer = styled.div`
     @media screen and (min-width: 768px) {
         width: 66%;
     }
