+++ client/components/ProductImage/ProductImage.styles.ts
@@ -1,0 +1,7 @@
+import styled from "styled-components"
+
+export const ProductImageContainer = styled.div`
+    @media screen and (max-width: 768px) {
+        text-align: center;
+    }
+`
